Fixed error in keylist management functions that masked changes to keys on other lists. The problem was that the predefined keylists were not being kept in sync and so were improperly reset when setting multiple lists. This affected `win-switch-custom-set-keys', `win-switch-add-key', and `win-switch-delete-key'. Thanks to Mark Hepburn and Brett Presnell for finding the error. Also fixed a few documentation typos.
As a note for the future, the separate keylist variables as the
functional store of the current configuraqtion are probably a bad idea
and should be replaced by a single data structure. The original idea was
to make customization a bit more intuitive, but perhaps a layer can be
added transparently to improve this.